% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvQuad.R
\name{copyNIGrid}
\alias{copyNIGrid}
\title{copies an NIGrid-object}
\usage{
copyNIGrid(object1, object2 = NULL)
}
\arguments{
\item{object1}{original NIGrid-object}

\item{object2}{destination; if NULL \code{copyNIGrid} returns a NIGrid-object
otherwise the \code{object2} will be overwritten.}
}
\value{
Returns a NIGrid-object or NULL
}
\description{
\code{copyNIGrid} copies an NIGrid-object
}
\examples{
myGrid <- createNIGrid(dim=2, type="GHe", level=5)
myGrid.copy <- copyNIGrid(myGrid)
}

