% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvQuad.R
\name{getNodes and getWeights}
\alias{getNodes}
\alias{getNodes and getWeights}
\alias{getWeights}
\title{get nodes and weights from an NIGrid-object}
\usage{
getNodes(grid)

getWeights(grid)
}
\arguments{
\item{grid}{object of class \code{NIGrid}}
}
\value{
Returns the nodes or weights of the given grid
}
\description{
\code{getNodes} and \code{getWeights} extract the (potentially rescaled) nodes and weights
 out of an NIGrid-Object
}
\examples{
myGrid <- createNIGrid(dim=2, type="cNC1", level=3)
getNodes(myGrid)
getWeights(myGrid)
}
\seealso{
\code{\link{createNIGrid}}
}

