% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvQuad.R
\name{plot (plot.NIGrid)}
\alias{plot (plot.NIGrid)}
\alias{plot.NIGrid}
\title{plots an NIGrid-object}
\usage{
\method{plot}{NIGrid}(x, plot.dimension = NULL, ...)
}
\arguments{
\item{x}{a grid of type \code{NIGrid}}

\item{plot.dimension}{vector of length 1, 2 or 3. with the dimensions to be plotted
(see examples)}

\item{...}{arguments passed to the default plot command}
}
\description{
Plots the grid points of an NIGrid-object
}
\examples{
myGrid <- createNIGrid(dim=4, type=c("GHe", "cNC1", "GLe", "oNC1"),
                       level=c(3,4,5,6))
plot(myGrid) ## dimension 1-min(3,dim(myGrid)) are plotted
## Free arranged plots
plot(myGrid, plot.dimension=c(4,2,1))
plot(myGrid, plot.dimension=c(1,2))
plot(myGrid, plot.dimension=c(3))
}

