% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvQuad.R
\name{print (print.NIGrid)}
\alias{print (print.NIGrid)}
\alias{print.NIGrid}
\title{prints characteristic information for an NIGrid-object}
\usage{
\method{print}{NIGrid}(x, ...)
}
\arguments{
\item{x}{a grid of type \code{NIGrid}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Prints the information for an NIGrid-object (i.a. grid size (dimensions, grid points, memory usage),
type and support)
}
\description{
Prints characteristic information for an NIGrid-object
}
\examples{
myGrid <- createNIGrid(dim=2, type="GHe", level=5)
print(myGrid)
}

